update codegen to scala 3
