+++ dnp3/codegen/src/main/scala/dev/gridio/dnp3/codegen/Main.scala
@@ -2,8 +2,11 @@ package dev.gridio.dnp3.codegen
 
 import java.nio.file.{FileSystems, Path}
 
-import dev.gridio.dnp3.codegen.render._
-import dev.gridio.dnp3.codegen.render.modules._
+import dev.gridio.dnp3.codegen.render.*
+import dev.gridio.dnp3.codegen.render.given
+import dev.gridio.dnp3.codegen.render.given_Conversion_String_RenderString
+import dev.gridio.dnp3.codegen.render.modules.*
+import dev.gridio.dnp3.codegen.writeTo
 
 object Main {
