+++ dnp3/codegen/src/main/scala/dev/gridio/dnp3/codegen/render/modules/ConversionsModule.scala
@@ -1,7 +1,8 @@
 package dev.gridio.dnp3.codegen.render.modules
 
 import dev.gridio.dnp3.codegen.model._
-import dev.gridio.dnp3.codegen.render._
+import dev.gridio.dnp3.codegen.render.*
+import dev.gridio.dnp3.codegen.render.given_Conversion_String_RenderString
 
 object ConversionsModule extends Module {
