+++ dnp3/codegen/src/main/scala/dev/gridio/dnp3/codegen/render/modules/EnumModule.scala
@@ -2,7 +2,8 @@ package dev.gridio.dnp3.codegen.render.modules
 
 import dev.gridio.dnp3.codegen.model._
 import dev.gridio.dnp3.codegen.model.enums.protocol._
-import dev.gridio.dnp3.codegen.render._
+import dev.gridio.dnp3.codegen.render.*
+import dev.gridio.dnp3.codegen.render.given_Conversion_String_RenderString
 
 object EnumModule {
   def application : Module = {
